fix: resolve CI failures for Hytte-wnep
